gpu: share the error latch the shader programs each copied
Both IR programs carried a byte-identical copy of the helper that latches the
first builder error so the program body can be written as a straight line of
expressions. A third program would have made three, so it moves to a header the
programs library exports and both existing copies are deleted.

Renamed from Latch to ErrorLatch on the way out: at file scope inside one
program the short name was unambiguous, but a shared header has to say what it
latches.
